pyproject.toml
[tool.poetry]
name = "scholar_bot"
version = "0.1.0"
description = ""
authors = ["Jeremy Keeler <j@keeler.dev>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
scholarly = "^0.4.1"
python-dotenv = "^0.14.0"
"discord.py" = "^1.4.1"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
